Allow PudlTabl to read data from Parquet outputs
To try and speed up our CI and potentially reduce memory usage, this
change adds a parameter to the PudlTabl output object allowing the user
to select whether the class should read from SQLite or Parquet outputs.
It uses the same method of reading from Parquet as our existing
ParquetIOManager in Dagster.

The new `table_source="parquet"` parameter is now used in all the CI
tests that actually read data.